+++ front/src/components/ScheduleTimeList/index.tsx
@@ -1,54 +1,22 @@
-import { useContext } from 'react';
-import { AxiosError } from 'axios';
-
-import { SnackbarContext } from '@context/SnackbarProvider';
 import { getHourMinutes } from '@utils/date';
-import { editCoachSchedule } from '@api/coach';
 import type { TimeSchedule } from '@typings/domain';
 import * as S from './styles';
 
 interface ScheduleTimeListProps {
-  date: string;
   isSelectedAll: boolean;
   daySchedule: TimeSchedule[];
   onClickTime: (dateTime: string) => void;
   onSelectAll: () => void;
-  onUpdateSchedule: (selectedTimes: string[]) => void;
+  onUpdateDaySchedule: () => Promise<void>;
 }
 
 const ScheduleTimeList = ({
-  date,
   isSelectedAll,
   daySchedule,
   onClickTime,
   onSelectAll,
-  onUpdateSchedule,
+  onUpdateDaySchedule,
 }: ScheduleTimeListProps) => {
-  const showSnackbar = useContext(SnackbarContext);
-
-  const getSelectedTimes = () => {
-    return daySchedule.reduce((newArray, { isSelected, dateTime }) => {
-      if (isSelected) {
-        newArray.push(dateTime);
-      }
-      return newArray;
-    }, [] as string[]);
-  };
-
-  const handleUpdateDaySchedule = async () => {
-    const selectedTimes = getSelectedTimes();
-    try {
-      await editCoachSchedule(date, selectedTimes);
-      onUpdateSchedule(selectedTimes);
-      showSnackbar({ message: '확정되었습니다. ✅' });
-    } catch (error) {
-      if (error instanceof AxiosError) {
-        alert(error.response?.data?.message);
-        console.log(error);
-      }
-    }
-  };
-
   return (
     <S.TimeListContainer>
       <S.ScrollContainer>
@@ -72,7 +40,7 @@ const ScheduleTimeList = ({
         <S.CheckButton onClick={onSelectAll}>
           {isSelectedAll ? '전체 해제' : '전체 선택'}
         </S.CheckButton>
-        <S.ConfirmButton onClick={handleUpdateDaySchedule}>확인</S.ConfirmButton>
+        <S.ConfirmButton onClick={onUpdateDaySchedule}>확인</S.ConfirmButton>
       </S.ButtonContainer>
     </S.TimeListContainer>
   );
