+++ front/src/pages/Reservation/index.tsx
@@ -1,20 +1,28 @@
 import { useEffect, useState } from 'react';
-import { useParams } from 'react-router-dom';
+import { useNavigate, useParams } from 'react-router-dom';
 import { AxiosError } from 'axios';
 
 import ReservationTimeList from '@components/ReservationTimeList';
 import Calendar from '@components/Calendar';
 import Frame from '@components/Frame';
 import Title from '@components/Title';
+import Modal from '@components/Modal';
 import useCalendar from '@hooks/useCalendar';
 import useBoolean from '@hooks/useBoolean';
 import { getCoachSchedulesByCrew } from '@api/coach';
+import { createReservation } from '@api/reservation';
+import { ROUTES } from '@constants/index';
 import type { DaySchedule, MonthScheduleMap, ScheduleInfo } from '@typings/domain';
 import theme from '@styles/theme';
 import * as S from '@styles/common';
 
+import CheckCircle from '@assets/check-circle.svg';
+
 const Reservation = () => {
+  const navigate = useNavigate();
   const { id: coachId } = useParams();
+  const { value: isOpenModal, setTrue: openModal, setFalse: closeModal } = useBoolean();
+  const [reservationId, setReservationId] = useState<number | null>(null);
   const { value: isOpenTimeList, setTrue: openTimeList, setFalse: closeTimeList } = useBoolean();
   const { monthYear, selectedDay, setSelectedDay, dateBoxLength, updateMonthYear } = useCalendar();
   const [selectedTimeId, setSelectedTimeId] = useState<number | null>(null);
@@ -48,11 +56,11 @@ const Reservation = () => {
     });
   };
 
-  const handleReservationTime = (scheduleId: number) => {
+  const updateTimeIsPossible = (scheduleId: number, isPossible: boolean) => {
     setSchedule((allSchedules) => {
       const newDaySchedule = schedule.daySchedule.map((time) => {
         if (time.id === scheduleId) {
-          return { id: scheduleId, dateTime: time.dateTime, isPossible: false };
+          return { id: scheduleId, dateTime: time.dateTime, isPossible };
         }
         return time;
       });
@@ -79,8 +87,24 @@ const Reservation = () => {
     setSelectedTimeId(null);
   };
 
-  const handleClickTime = (id: number | null) => {
+  const handleClickTime = (id: number) => {
     setSelectedTimeId(id);
+  };
+
+  const handleClickReservation = async (scheduleId: number) => {
+    try {
+      const data = await createReservation(scheduleId);
+      const location = data.headers.location.split('/').pop();
+      setReservationId(Number(location));
+      setSelectedTimeId(null);
+      updateTimeIsPossible(scheduleId, false);
+      openModal();
+    } catch (error) {
+      if (error instanceof AxiosError) {
+        alert(error.response?.data?.message);
+        console.log(error);
+      }
+    }
   };
 
   useEffect(() => {
@@ -122,13 +146,27 @@ const Reservation = () => {
           {isOpenTimeList && (
             <ReservationTimeList
               daySchedule={schedule.daySchedule}
-              onReservationTime={handleReservationTime}
               selectedTimeId={selectedTimeId}
               onClickTime={handleClickTime}
+              onClickReservation={handleClickReservation}
             />
           )}
         </S.CalendarContainer>
       </S.ScheduleContainer>
+
+      {isOpenModal && (
+        <Modal
+          icon={CheckCircle}
+          title="예약완료"
+          firstButtonName="나중에"
+          secondButtonName="작성하기"
+          onClickFirstButton={() => navigate(ROUTES.CREW_HISTORY)}
+          onClickSecondButton={() => navigate(`${ROUTES.CREW_SHEET}/${reservationId}`)}
+          closeModal={closeModal}
+        >
+          <p>면담 내용을 지금 작성 하시겠습니까?</p>
+        </Modal>
+      )}
     </Frame>
   );
 };
